fix: update watcher with api fixes (#30)
* Fix: Update watcher with api fixes

There are two primary things that this does:

1. This fixes the redirect url problem. It changed from having localhost
be acceptable to needing it to be a loopback url. See
https://developer.spotify.com/documentation/web-api/concepts/redirect_uri

2. The `audio_features` api is discontinued by spotify for new webapps.
Having the method call be unwrapped with a try except block killed
the watcher. This change just wraps the audio_features in a try except
block so that people who have old apps that currently have access to
the API route can still use it. See below for info about changes to
the Spotify Web API:
https://developer.spotify.com/blog/2024-11-27-changes-to-the-web-api

* Fix: use pure oath for auth

* Fix: change to AW based logging

Errors were not logged with the original type of logging. This
makes the errors or other logs go to the default AW logging places.

* Fix: Correct redirect uri

* Fix: Move args into main

---------

Co-authored-by: Erik Bjäreholt <erik@bjareho.lt>